factor out data
